backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pandas==2.1.4
openpyxl==3.1.2
bcrypt==4.1.2
pyjwt==2.8.0
reportlab==4.0.9
gunicorn==21.2.0
